app/requirements.txt
fastapi
uvicorn
pydantic
torch
torchvision
mlflow
Pillow
pytest
requests
h5py